refactor(hir): extract computed-member helpers to class_computed.rs (file-size gate)
